pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mlpipeline"
version = "0.1"
description = "mlpipeline does something"
readme = "README.md"
license = {file = "LICENSE"}
authors = [
    {name = "VNPT@IC"}
]
requires-python = ">=3.11"
dependencies = [
    "pandas==2.2.3",
    "PyYAML==6.0.2",
    "pydantic>=2.0.0",
    "scikit-learn==1.5.2",
    "loguru==0.7.3",
]

[project.optional-dependencies]
dev = [
    "pandas==2.2.3",
    "black==23.3.0",
    "flake8==6.0.0",
    "mypy==1.8.0",
    "isort==5.12.0",
    "types-PyYAML==6.0.12.10",
    "pre-commit==3.3.2",
    "pylint==3.2.7",
    "pytest==7.3.2"
]
test = [
    "pandas==2.2.3",
    "pytest==7.3.2"
]

[project.scripts]
mlpipeline = "mlpipeline.__main__:main"

[tool.setuptools.packages.find]
exclude = ["tests", "docs"]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
